Reorganize aries-vcx libindy module
