+++ aries_vcx/src/global/settings.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 use std::sync::RwLock;
 
 use crate::error::prelude::*;
-use crate::libindy::utils::wallet::IssuerConfig;
+use crate::libindy::wallet::IssuerConfig;
 
 pub static CONFIG_POOL_NAME: &str = "pool_name";
 pub static CONFIG_SDK_TO_REMOTE_ROLE: &str = "sdk_to_remote_role";
