+++ aries_vcx/src/libindy/crypto.rs
