+++ aries_vcx/src/libindy/primitives/mod.rs
@@ -1,0 +1,3 @@
+pub mod revocation_registry;
+pub mod credential_definition;
+pub mod credential_schema;
