+++ aries_vcx/src/libindy/proofs/proof_request.rs
@@ -5,7 +5,7 @@ use serde_json;
 
 use crate::error::prelude::*;
 use crate::libindy::proofs::proof_request_internal::{AttrInfo, NonRevokedInterval, PredicateInfo};
-use crate::libindy::utils::anoncreds;
+use crate::libindy::anoncreds;
 
 #[derive(Serialize, Deserialize, Builder, Debug, PartialEq, Clone)]
 #[builder(setter(into), default)]
