+++ aries_vcx/src/libindy/proofs/verifier/verifier_internal.rs
@@ -4,7 +4,7 @@ use serde_json::Value;
 
 use crate::error::prelude::*;
 use crate::global::settings;
-use crate::libindy::utils::anoncreds;
+use crate::libindy::anoncreds;
 use crate::utils::openssl::encode;
 
 #[derive(Debug, Deserialize, Serialize, PartialEq)]
