+++ aries_vcx/src/libindy/utils/mod.rs
@@ -5,15 +5,8 @@ use vdrtools_sys::CommandHandle;
 
 use crate::global::settings;
 
-pub mod anoncreds;
-pub mod cache;
-pub mod crypto;
-pub mod ledger;
 pub mod logger;
 pub mod mocks;
-pub mod pool;
-pub mod signus;
-pub mod wallet;
 
 pub mod error_codes;
