+++ aries_vcx/src/protocols/issuance/issuer/state_machine.rs
@@ -5,7 +5,7 @@ use vdrtools_sys::WalletHandle;
 
 use crate::error::{VcxError, VcxErrorKind, VcxResult};
 use crate::libindy::credentials::encode_attributes;
-use crate::libindy::utils::anoncreds;
+use crate::libindy::anoncreds;
 use messages::a2a::{A2AMessage, MessageId};
 use messages::issuance::credential::Credential;
 use messages::issuance::credential_offer::{CredentialOffer, OfferInfo};
