+++ aries_vcx/tests/test_connection.rs
@@ -54,7 +54,7 @@ mod integration_tests {
     #[tokio::test]
     async fn test_oob_connection_bootstrap() {
         use messages::connection::invite::Invitation;
-        use aries_vcx::libindy::utils::ledger::into_did_doc;
+        use aries_vcx::libindy::ledger::transactions::into_did_doc;
         let setup = SetupPool::init().await;
         let mut institution = Faber::setup(setup.pool_handle).await;
         let mut consumer = Alice::setup(setup.pool_handle).await;
