+++ aries_vcx/tests/test_creds_proofs.rs
@@ -11,8 +11,8 @@ pub mod utils;
 #[cfg(feature = "pool_tests")]
 mod integration_tests {
     use aries_vcx::handlers::proof_presentation::prover::Prover;
-    use aries_vcx::libindy::utils::anoncreds::get_cred_def_json;
-    use aries_vcx::libindy::utils::anoncreds::test_utils::{
+    use aries_vcx::libindy::anoncreds::get_cred_def_json;
+    use aries_vcx::libindy::anoncreds::test_utils::{
         create_and_store_credential, create_and_store_nonrevocable_credential,
         create_and_store_nonrevocable_credential_def, create_indy_proof,
     };
@@ -321,7 +321,8 @@ mod tests {
     use aries_vcx::handlers::proof_presentation::prover::Prover;
     use aries_vcx::handlers::proof_presentation::verifier::Verifier;
     use aries_vcx::libindy;
-    use aries_vcx::libindy::utils::anoncreds::test_utils::create_and_store_nonrevocable_credential_def;
+    use aries_vcx::libindy::anoncreds::test_utils::create_and_store_nonrevocable_credential_def;
+    use aries_vcx::libindy::ledger::pool::test_utils::{delete_test_pool, open_test_pool};
     use aries_vcx::messages::issuance::credential_offer::CredentialOffer;
     use aries_vcx::messages::proof_presentation::presentation_request::PresentationRequest;
     use aries_vcx::protocols::issuance::holder::state_machine::HolderState;
@@ -1138,14 +1139,14 @@ mod tests {
 
     impl Pool {
         pub async fn open() -> Pool {
-            let handle = libindy::utils::pool::test_utils::open_test_pool().await;
+            let handle = open_test_pool().await;
             Pool { handle }
         }
     }
 
     impl Drop for Pool {
         fn drop(&mut self) {
-            futures::executor::block_on(libindy::utils::pool::test_utils::delete_test_pool(self.handle));
+            futures::executor::block_on(delete_test_pool(self.handle));
         }
     }
