+++ aries_vcx/tests/test_mysql_wallet.rs
@@ -37,10 +37,8 @@ mod dbtests {
     use agency_client::configuration::AgentProvisionConfig;
     use aries_vcx::global::settings;
     use aries_vcx::global::settings::init_issuer_config;
-    use aries_vcx::libindy::utils::wallet::{
-        close_wallet, create_wallet_with_master_secret, wallet_configure_issuer, WalletConfig, WalletConfigBuilder,
-    };
-    use aries_vcx::libindy::wallet::open_wallet;
+    use aries_vcx::libindy::wallet::WalletConfigBuilder;
+    use aries_vcx::libindy::wallet::{close_wallet, create_wallet_with_master_secret, open_wallet, wallet_configure_issuer, WalletConfig};
     use aries_vcx::utils::devsetup::{AGENCY_DID, AGENCY_ENDPOINT, AGENCY_VERKEY};
     use aries_vcx::utils::provision::provision_cloud_agent;
     use aries_vcx::utils::test_logger::LibvcxDefaultLogger;
