+++ aries_vcx/tests/utils/scenarios.rs
@@ -16,11 +16,11 @@ pub mod test_utils {
     use aries_vcx::handlers::proof_presentation::prover::Prover;
     use aries_vcx::handlers::proof_presentation::verifier::Verifier;
     use aries_vcx::libindy;
-    use aries_vcx::libindy::credential_def::revocation_registry::RevocationRegistry;
-    use aries_vcx::libindy::credential_def::CredentialDef;
-    use aries_vcx::libindy::utils::anoncreds::publish_local_revocations;
-    use aries_vcx::libindy::utils::anoncreds::test_utils::create_and_store_credential_def;
-    use aries_vcx::libindy::utils::ledger::into_did_doc;
+    use aries_vcx::libindy::primitives::revocation_registry::RevocationRegistry;
+    use aries_vcx::libindy::primitives::credential_definition::CredentialDef;
+    use aries_vcx::libindy::anoncreds::publish_local_revocations;
+    use aries_vcx::libindy::anoncreds::test_utils::create_and_store_credential_def;
+    use aries_vcx::libindy::ledger::transactions::into_did_doc;
     use aries_vcx::messages::connection::invite::Invitation;
     use aries_vcx::messages::issuance::credential_offer::{CredentialOffer, OfferInfo};
     use aries_vcx::messages::issuance::credential_proposal::{CredentialProposal, CredentialProposalData};
@@ -696,7 +696,7 @@ pub mod test_utils {
     }
 
     pub async fn revoke_credential_local(faber: &mut Faber, issuer_credential: &Issuer, rev_reg_id: &str) {
-        let (_, delta, timestamp) = libindy::utils::anoncreds::get_rev_reg_delta_json(faber.pool_handle, &rev_reg_id, None, None)
+        let (_, delta, timestamp) = libindy::anoncreds::get_rev_reg_delta_json(faber.pool_handle, &rev_reg_id, None, None)
             .await
             .unwrap();
         info!("revoking credential locally");
@@ -705,7 +705,7 @@ pub mod test_utils {
             .await
             .unwrap();
         let (_, delta_after_revoke, _) =
-            libindy::utils::anoncreds::get_rev_reg_delta_json(faber.pool_handle, rev_reg_id, Some(timestamp + 1), None)
+            libindy::anoncreds::get_rev_reg_delta_json(faber.pool_handle, rev_reg_id, Some(timestamp + 1), None)
                 .await
                 .unwrap();
         assert_ne!(delta, delta_after_revoke); // They will not equal as we have saved the delta in cache
@@ -734,7 +734,7 @@ pub mod test_utils {
     }
 
     pub async fn publish_revocation(institution: &mut Faber, rev_reg_id: String) {
-        libindy::utils::anoncreds::publish_local_revocations(institution.wallet_handle, institution.pool_handle, &institution.config_issuer.institution_did, rev_reg_id.as_str())
+        libindy::anoncreds::publish_local_revocations(institution.wallet_handle, institution.pool_handle, &institution.config_issuer.institution_did, rev_reg_id.as_str())
             .await
             .unwrap();
     }
