+++ libvcx/src/api_lib/api_c/utils.rs
@@ -424,7 +424,7 @@ pub extern "C" fn vcx_endorse_transaction(
     );
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::ledger::endorse_transaction(get_main_wallet_handle(), pool_handle, &issuer_did, &transaction).await {
+        match aries_vcx::libindy::ledger::transactions::endorse_transaction(get_main_wallet_handle(), pool_handle, &issuer_did, &transaction).await {
             Ok(()) => {
                 trace!(
                     "vcx_endorse_transaction(command_handle: {}, issuer_did: {}, rc: {})",
@@ -469,7 +469,7 @@ pub extern "C" fn vcx_rotate_verkey(
     };
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::signus::rotate_verkey(get_main_wallet_handle(), pool_handle, &did).await {
+        match aries_vcx::libindy::signus::rotate_verkey(get_main_wallet_handle(), pool_handle, &did).await {
             Ok(()) => {
                 trace!(
                     "vcx_rotate_verkey_cb(command_handle: {}, rc: {})",
@@ -508,7 +508,7 @@ pub extern "C" fn vcx_rotate_verkey_start(
     );
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::signus::libindy_replace_keys_start(get_main_wallet_handle(), &did).await {
+        match aries_vcx::libindy::signus::libindy_replace_keys_start(get_main_wallet_handle(), &did).await {
             Ok(temp_vk) => {
                 trace!(
                     "vcx_rotate_verkey_start_cb(command_handle: {}, rc: {}, temp_vk: {})",
@@ -560,7 +560,7 @@ pub extern "C" fn vcx_rotate_verkey_apply(
     };
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::signus::rotate_verkey_apply(get_main_wallet_handle(), pool_handle, &did, &temp_vk).await {
+        match aries_vcx::libindy::signus::rotate_verkey_apply(get_main_wallet_handle(), pool_handle, &did, &temp_vk).await {
             Ok(()) => {
                 trace!(
                     "vcx_rotate_verkey_apply_cb(command_handle: {}, rc: {})",
@@ -602,7 +602,7 @@ pub extern "C" fn vcx_get_verkey_from_wallet(
     );
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::signus::get_verkey_from_wallet(get_main_wallet_handle(), &did).await {
+        match aries_vcx::libindy::signus::get_verkey_from_wallet(get_main_wallet_handle(), &did).await {
             Ok(verkey) => {
                 trace!(
                     "vcx_get_verkey_from_wallet_cb(command_handle: {}, rc: {}, verkey: {})",
@@ -651,7 +651,7 @@ pub extern "C" fn vcx_get_verkey_from_ledger(
     };
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::signus::get_verkey_from_ledger(pool_handle, &did).await {
+        match aries_vcx::libindy::signus::get_verkey_from_ledger(pool_handle, &did).await {
             Ok(verkey) => {
                 trace!(
                     "vcx_get_verkey_from_ledger_cb(command_handle: {}, rc: {}, verkey: {})",
@@ -701,7 +701,7 @@ pub extern "C" fn vcx_get_ledger_txn(
     };
 
     execute_async::<BoxFuture<'static, Result<(), ()>>>(Box::pin(async move {
-        match aries_vcx::libindy::utils::anoncreds::get_ledger_txn(
+        match aries_vcx::libindy::anoncreds::get_ledger_txn(
             get_main_wallet_handle(),
             pool_handle,
             submitter_did.as_deref(),
