+++ libvcx/src/api_lib/global/pool.rs
@@ -2,8 +2,8 @@ use std::sync::RwLock;
 
 use aries_vcx::error::{VcxError, VcxErrorKind, VcxResult};
 use aries_vcx::global::settings;
-use aries_vcx::libindy::utils::pool::PoolConfig;
-use aries_vcx::libindy::utils::pool::{create_pool_ledger_config, open_pool_ledger, close};
+use aries_vcx::libindy::ledger::pool::PoolConfig;
+use aries_vcx::libindy::ledger::pool::{create_pool_ledger_config, open_pool_ledger, close};
 use aries_vcx::vdrtools::INVALID_POOL_HANDLE;
 
 lazy_static! {
