+++ libvcx/src/api_lib/global/wallet.rs
@@ -1,9 +1,9 @@
 use aries_vcx::error::VcxResult;
 use aries_vcx::global::settings;
-use aries_vcx::vdrtools::{WalletHandle, INVALID_WALLET_HANDLE};
+use aries_vcx::vdrtools::{INVALID_WALLET_HANDLE, WalletHandle};
 use aries_vcx::libindy;
-use aries_vcx::libindy::utils::wallet::WalletConfig;
-use aries_vcx::libindy::utils::{anoncreds, wallet};
+use aries_vcx::libindy::wallet::WalletConfig;
+use aries_vcx::libindy::{anoncreds, wallet};
 
 pub static mut WALLET_HANDLE: WalletHandle = INVALID_WALLET_HANDLE;
 
@@ -25,7 +25,7 @@ pub fn reset_main_wallet_handle() -> VcxResult<()> {
 }
 
 pub async fn export_main_wallet(path: &str, backup_key: &str) -> VcxResult<()> {
-    wallet::export_wallet(get_main_wallet_handle(), path, backup_key).await
+    libindy::wallet::export_wallet(get_main_wallet_handle(), path, backup_key).await
 }
 
 pub async fn open_as_main_wallet(wallet_config: &WalletConfig) -> VcxResult<WalletHandle> {
@@ -35,13 +35,13 @@ pub async fn open_as_main_wallet(wallet_config: &WalletConfig) -> VcxResult<Wall
 }
 
 pub async fn create_and_open_as_main_wallet(wallet_config: &WalletConfig) -> VcxResult<WalletHandle> {
-    let handle = wallet::create_and_open_wallet(wallet_config).await?;
+    let handle = libindy::wallet::create_and_open_wallet(wallet_config).await?;
     set_main_wallet_handle(handle);
     Ok(handle)
 }
 
 pub async fn close_main_wallet() -> VcxResult<()> {
-    wallet::close_wallet(get_main_wallet_handle()).await?;
+    libindy::wallet::close_wallet(get_main_wallet_handle()).await?;
     reset_main_wallet_handle()?;
     Ok(())
 }
@@ -63,9 +63,9 @@ pub async fn create_main_wallet(config: &WalletConfig) -> VcxResult<()> {
 pub mod test_utils {
     use aries_vcx::global;
     use aries_vcx::global::settings;
-    use aries_vcx::libindy::utils::signus::create_and_store_my_did;
-    use aries_vcx::libindy::utils::wallet::add_wallet_record;
-    use aries_vcx::libindy::utils::wallet::*;
+    use aries_vcx::libindy::signus::create_and_store_my_did;
+    use aries_vcx::libindy::wallet::{add_wallet_record, WalletConfig};
+    use aries_vcx::libindy::wallet::*;
     use aries_vcx::utils::devsetup::TempFile;
 
     use crate::api_lib::global::wallet::{close_main_wallet, create_and_open_as_main_wallet, export_main_wallet};
